@@ -1,12 +1,12 @@
 import 'dart:io';
-import 'package:path/path.dart';
+
 import 'package:html/dom.dart';
+import 'package:path/path.dart';
 import 'package:yacht/src/html_printer.dart';
 import 'package:yacht/src/html_visitor.dart';
-import 'package:meta/meta.dart';
 
 /// Tidy a file, if [dstFilePath] is ommited, original file is replaced
-Future tidyHtml({@required String srcFilePath, String dstFilePath}) async {
+Future tidyHtml({required String srcFilePath, String? dstFilePath}) async {
   dstFilePath ??= srcFilePath;
   var src = await File(srcFilePath).readAsString();
   var builder = HtmlDocumentVisitor();
