@@ -29,7 +29,7 @@ class HtmlDocumentVisitor extends HtmlElementVisitor {
 
   // public API to call
   Document visitDocument(Document document) {
-    void _visitElement(Element element) {
+    void _visitElement(Element? element) {
       if (element != null) {
         visitElement(element);
       }
